Complete refactor of XA1DE & XBAODE with new ExpClf class
